+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bug #3: Fix trait `HasListContainerCollection` class and update `README.md` (@terabytesoftw)
 - Enh #4: Add trait `HasFirstItemClass` class (@terabytesoftw)
 - Enh #5: Add trait `HasFirstLinkClass` class (@terabytesoftw)
+- Enh #6: Add trait `HasLastItemClass` class (@terabytesoftw)
 
 ## 0.1.0 March 5, 2024